SPOCAN-59 agrego javadocs y refactorizo nombres de paquetes
