+++ src/main/java/com/spochi/auth/AuthorizationException.java
@@ -1,3 +1,3 @@
-package com.spochi.auth.firebase;
+package com.spochi.auth;
 
 public class AuthorizationException extends RuntimeException {}
