+++ src/test/java/com/spochi/auth/JwtFilterTest.java
@@ -1,6 +1,6 @@
 package com.spochi.auth;
 
-import com.spochi.service.authenticate.JwtUtil;
+import com.spochi.service.auth.JwtUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
